correções

+++ Exercícios/007 - media.py
@@ -1,4 +1,4 @@
-nota1 = int(input('digite a primeira nota do aluno: '))
-nota2 = int(input('Digite a segunda nota do aluno: '))
+nota1 = float(input('digite a primeira nota do aluno: '))
+nota2 = float(input('Digite a segunda nota do aluno: '))
 media = (nota1 + nota2) / 2
 print(f'A média das notas foi de {media}')
